requirements.txt
matplotlib~=3.8.0
requests~=2.31.0
Pillow~=9.5.0
chardet~=5.2.0
beautifulsoup4~=4.12.2
selenium~=4.14.0
ddddocr~=1.4.8